Applied server timing filter
